abstracting tabs section
